Fix error occuring when invalid JSON received
Hi !

I got an error when working on a project with Laravel Telescope running. It try to execute a big query to insert data in `telescope_entries` table and the JSON data received was invalid (truncated) causing `$query` to be null.

This PR is a proposition to fix this without ignoring that a query was run.

Have a nice day! 😃